feat : 수정하기, 질문삭제, 답변거절 기능 추가
rebase를 제대로 끝내지 않고 작업을 진행,
머지 커밋된 것의 메세지를  수정한 것입니다.
수정할때 깃모지는 어떻게 넣는지 몰라 넣지 않았습니다.
