+++ src/components/Button/BadgeButton/BadgeButtonStyle.js
@@ -1,5 +1,10 @@
 import styled from 'styled-components';
 
+const COLORS = {
+  brown: 'var(--Brown-40)',
+  gray: 'var(--Grayscale-40)',
+};
+
 export const Button = styled.button`
   display: inline-flex;
   padding: 4px 12px;
@@ -7,8 +12,7 @@ export const Button = styled.button`
   align-items: center;
   gap: 10px;
 
-  color: ${({ $isAnswered }) =>
-    $isAnswered ? 'var(--Brown-40)' : 'var(--Grayscale-40)'};
+  color: ${({ $color }) => COLORS[$color]};
   font-feature-settings:
     'clig' off,
     'liga' off;
@@ -17,8 +21,6 @@ export const Button = styled.button`
   line-height: 18px; /* 128.571% */
 
   border-radius: 8px;
-  border: 1px solid
-    ${({ $isAnswered }) =>
-      $isAnswered ? 'var(--Brown-40)' : 'var(--Grayscale-40)'};
+  border: 1px solid ${({ $color }) => COLORS[$color]};
   background: var(--Grayscale-10);
 `;
