+++ src/components/Button/KebabButton/KebabButton.jsx
@@ -1,10 +1,7 @@
 import { useState } from 'react';
-import { ReactComponent as Edit } from '../../images/Edit.svg';
-import { ReactComponent as Rejection } from '../../images/Rejection.svg';
-import { ReactComponent as Close } from '../../images/Close.svg';
 import * as S from './KebabButtonStyle';
 
-export default function Kebab() {
+export default function Kebab({ menuItem, questionId, question }) {
   const [isOpenKebabMenu, setIsOpenKebabMenu] = useState();
   const [selectedMenuItem, setSelectedMenuItem] = useState();
 
@@ -12,44 +9,27 @@ export default function Kebab() {
     setIsOpenKebabMenu(!isOpenKebabMenu);
   };
 
-  const handleKebabMenuItemOnClick = e => {
-    setSelectedMenuItem(e.target.innerText);
+  const handleKebabButtonOnBlur = () => {
+    setTimeout(() => {
+      setIsOpenKebabMenu(false);
+    }, 200);
   };
 
-  const menuItem = [
-    {
-      text: '수정하기',
-      imagePath: <Edit fill="" />,
-      imageBluePath: <Edit fill="var(--Blue-50)" />,
-      imageAlt: '수정하기 아이콘',
-    },
-    {
-      text: '질문삭제',
-      imagePath: <Close fill="" />,
-      imageBluePath: <Close fill="var(--Blue-50)" />,
-      imageAlt: '질문삭제 아이콘',
-    },
-    {
-      text: '답변거절',
-      imagePath: <Rejection fill="" />,
-      imageBluePath: <Rejection fill="var(--Blue-50)" />,
-      imageAlt: '답변거절 아이콘',
-    },
-  ];
-  const menuItemClassName = [];
-
-  menuItem.forEach(element => {
-    if (element === selectedMenuItem) {
-      menuItemClassName.push('selected');
+  const handleKebabMenuItemOnClick = e => {
+    if (e.target.innerText === selectedMenuItem) {
+      setSelectedMenuItem(null);
     } else {
-      menuItemClassName.push('');
+      setSelectedMenuItem(e.target.innerText);
     }
-  });
+  };
 
   return (
     <S.KebabContainer>
-      <S.KebabButton onClick={handleKebabButtonOnClick}>
-        <img src="./images/More.png" alt="케밥 이미지" />
+      <S.KebabButton
+        onClick={handleKebabButtonOnClick}
+        onBlur={handleKebabButtonOnBlur}
+      >
+        <img src="/images/More.png" alt="케밥 이미지" />
       </S.KebabButton>
       {isOpenKebabMenu && (
         <S.KebabMenu>
@@ -60,10 +40,15 @@ export default function Kebab() {
               className = 'selected';
               image = element.imageBluePath;
             }
+
             return (
               <S.KebabMenuItem
+                key={element.text}
                 className={className}
-                onClick={handleKebabMenuItemOnClick}
+                onClick={e => {
+                  handleKebabMenuItemOnClick(e);
+                  element.onClick(questionId);
+                }}
               >
                 {image}
                 <span>{element.text}</span>
